Rename src/components/header to src/components/timeline
