+++ src/components/timeline/EmptyThreadIndicator.css
@@ -2,14 +2,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-.headerEmptyThreadIndicator {
+.timelineEmptyThreadIndicator {
   --empty-buffer-color1: rgba(69, 160, 255, 0.225);
   --empty-buffer-color2: rgba(69, 160, 255, 0.075);
   /* The following color is based off grey-20, but uses opacity instead. */
   --shutdown-startup-color: rgba(0, 0, 50, 0.07);
 }
 
-.headerEmptyThreadIndicatorBlock {
+.timelineEmptyThreadIndicatorBlock {
   position: absolute;
   top: 0;
   bottom: 0;
@@ -17,19 +17,19 @@
   box-sizing: border-box;
 }
 
-.headerEmptyThreadIndicatorStartup {
+.timelineEmptyThreadIndicatorStartup {
   border-right: solid 1px var(--grey-30);
   background-color: var(--shutdown-startup-color);
   box-shadow: inset -4px 0 4px rgba(0, 0, 0, 0.05);
 }
 
-.headerEmptyThreadIndicatorShutdown {
+.timelineEmptyThreadIndicatorShutdown {
   border-left: solid 1px var(--grey-30);
   background-color: var(--shutdown-startup-color);
   box-shadow: inset 4px 0 4px rgba(0, 0, 0, 0.05);
 }
 
-.headerEmptyThreadIndicatorEmptyBuffer {
+.timelineEmptyThreadIndicatorEmptyBuffer {
   background: repeating-linear-gradient(
     45deg,
     var(--empty-buffer-color1),
@@ -39,7 +39,7 @@
   );
 }
 
-.headerEmptyThreadIndicatorLongTooltip {
+.timelineEmptyThreadIndicatorLongTooltip {
   display: inline-block;
   max-width: 500px;
 }
