+++ src/components/timeline/SelectionScrubberOverlay.js
