+++ src/components/timeline/ThreadStackGraph.js
