+++ src/test/components/EmptyThreadIndicator.test.js
@@ -11,7 +11,7 @@ import renderer from 'react-test-renderer';
 
 import EmptyThreadIndicator, {
   getIndicatorPositions,
-} from '../../components/header/EmptyThreadIndicator';
+} from '../../components/timeline/EmptyThreadIndicator';
 import { getProfileFromTextSamples } from '../fixtures/profiles/make-profile';
 import { getBoundingBox } from '../fixtures/utils';
 import mockRaf from '../fixtures/mocks/request-animation-frame';
